feat(web/Spaces): Sidebar with Add To Space functionality (#7602)
* fix: add mock providers for the Safe Sidebar so all NavItems get rendered

* fix: UI adjustments for the Spaces Selector Dropdown

* feat: Add To Space back btn + Spaces dropdown UI in Stories

* feat: new sidebar on the Safe level (WIP)

* refactor: move new hooks and components to separate files; move test files to corresponding folders

* fix: Storybook fixes for stories with Back To Space btn

* refactor: remove unused CSS classes

* refactor: separate spaceId extraction fn in utils

* feat: notification on success when Safe is added to Space

* fix: treat a Safe that is added to a Space as a Qualified Safe

* refactor: wrap sidebar functions in useCallback

* fix: collapsed API button in the Spaces Sidebar (#7609)

* feat: New Transaction btn

* test: add tests for the new New Tx Btn

* fix: UI fixes for sidebar elements

* feat: disable adding to a Space if Safes per Space limit is exceeded

* refactor: minior refactoring of Space limit logic

* test: add tests for the Space limit functionality

* fix: typing

* feat: no Add To Space btn for cf safes + tests

* feat: stories for different Safe Sidebar states

* feat: add tracking events to the new sidebar + tests

* feat: add Add Spaces popup modal component + story

* refactor: extract the workspace header logic to a new component; add Add To Space modal to the UI

* feat: add Shadcn Dialog and wrap App To Space popup into it

* fix: UI fixes for the Add To Space modal

* fix: add hydration check to the Safe sidebar

* test: improve test coverage for new components

* refactor: adjust z-indexes

* fix: granular error notifications; minor UI fixes

* refactor: move the Add To Space modal into a separate spaces component

* refactor: rectructure Sidebar components

* refactor: simplify Space selector dropdown for better readability

* test: increase test coverage

* fix: add guard to preventback to space navigation if spaceId is falsy; refactor: minor refactoring

* fix: skip failing smoke tests (will be adjusted in another PR)

* refactor: simplify components complexity; address issues; skip failing tests

* fix: UI: sidebar icons styles match header icons styles

* feat: wrap sidebar icons in a tooltip

* refactor: prevent a re-render of nav items when tx data arrives later

* test: fix failing tests

* fix: remove duplicated closing but from Add To Space modal

* fix: show API errors in the toaster on safe addition

* chore: remove unused css classes

* refactor: remove redundant spaceName prop, derive from selectedSpace.name directly

* refactor: rename `NewTransactionButton` to `SidebarActionButton` as it handles not only new txs

* feat: highlight active item icon in the sidebar

* test: add more tests for the safe sidebar
