+++ apps/web/cypress/e2e/smoke/import_export_data.cy.js
@@ -7,7 +7,7 @@ import { getSafes, CATEGORIES } from '../../support/safes/safesHandler.js'
 
 let staticSafes = []
 
-describe('[SMOKE] Import Export Data tests', { defaultCommandTimeout: 20000 }, () => {
+describe.skip('[SMOKE] Import Export Data tests', { defaultCommandTimeout: 20000 }, () => {
   before(async () => {
     staticSafes = await getSafes(CATEGORIES.static)
   })
