+++ apps/web/src/features/spaces/components/Sidebar/SidebarCommonFooter/SidebarCommonFooter.tsx
@@ -1,56 +1,31 @@
-import { useState, useCallback, type ReactElement } from 'react'
-import { Sparkles } from 'lucide-react'
+import type { ReactElement } from 'react'
 import { SidebarFooter, SidebarMenu, SidebarMenuItem, SidebarMenuButton } from '@/components/ui/sidebar'
 import { cn } from '@/utils/cn'
-import { icons } from '../config'
-import css from '../styles.module.css'
+import { icons } from './config'
+import css from './styles.module.css'
 import { IS_PRODUCTION } from '@/config/constants'
-import { trackEvent, OVERVIEW_EVENTS, MixpanelEventParams } from '@/services/analytics'
+import { HELP_CENTER_URL } from '@safe-global/utils/config/constants'
 import { Switch } from '@/components/ui/switch'
 import { Field, FieldLabel } from '@/components/ui/field'
 import { setDarkMode } from '@/store/settingsSlice'
 import { useDarkMode } from '@/hooks/useDarkMode'
-import { useAppDispatch, useAppSelector } from '@/store'
-import { CookieAndTermType, hasConsentFor } from '@/store/cookiesAndTermsSlice'
-import { openCookieBanner } from '@/store/popupSlice'
-import { BEAMER_SELECTOR } from '@/services/beamer'
-import { ApiCtaSidebar } from '../ApiCtaSidebar'
-import { SidebarIndexingStatus } from '../SidebarIndexingStatus'
+import { useAppDispatch } from '@/store'
+import { ApiCtaSidebar } from './ApiCtaSidebar'
 import useLocalStorage from '@/services/local-storage/useLocalStorage'
 import { LS_KEY } from '@/config/gateway'
-import HelpMenu from '@/components/common/HelpMenu'
-import { Tooltip, TooltipContent, TooltipTrigger } from '@/components/ui/tooltip'
 
 export const SidebarCommonFooter = ({ isSafeSidebar = false }: { isSafeSidebar?: boolean }): ReactElement => {
   const dispatch = useAppDispatch()
-  const hasBeamerConsent = useAppSelector((state) => hasConsentFor(state, CookieAndTermType.UPDATES))
   const isDarkMode = useDarkMode()
   const [isProdGateway = false, setIsProdGateway] = useLocalStorage<boolean>(LS_KEY)
-  const [helpMenuAnchor, setHelpMenuAnchor] = useState<HTMLElement | null>(null)
 
   const onToggleGateway = (checked: boolean) => {
     setIsProdGateway(checked)
     setTimeout(() => location.reload(), 300)
   }
 
-  const handleHelpClick = useCallback((event: React.MouseEvent<HTMLButtonElement>) => {
-    trackEvent({ ...OVERVIEW_EVENTS.HELP_CENTER }, { [MixpanelEventParams.SIDEBAR_ELEMENT]: 'Help Center' })
-    setHelpMenuAnchor(event.currentTarget)
-  }, [])
-
-  const handleHelpMenuClose = useCallback(() => {
-    setHelpMenuAnchor(null)
-  }, [])
-
-  const handleBeamerClick = useCallback(() => {
-    trackEvent({ ...OVERVIEW_EVENTS.WHATS_NEW }, { [MixpanelEventParams.SIDEBAR_ELEMENT]: "What's New" })
-    if (!hasBeamerConsent) {
-      dispatch(openCookieBanner({ warningKey: CookieAndTermType.UPDATES }))
-    }
-  }, [dispatch, hasBeamerConsent])
-
   return (
-    <SidebarFooter data-testid="sidebar-common-footer">
+    <SidebarFooter data-testid="sidebar-common-footer" className={css.sidebarFooter}>
       {/* Dev Toggles - only in non-production */}
       {!IS_PRODUCTION && (
         <div className="flex flex-col gap-2 px-3 py-2 group-data-[collapsible=icon]:hidden">
@@ -71,14 +46,16 @@ export const SidebarCommonFooter = ({ isSafeSidebar = false }: { isSafeSidebar?:
         </div>
       )}
 
-      <SidebarMenu className="gap-0.5">
-        <ApiCtaSidebar />
+      <ApiCtaSidebar />
 
-        <SidebarMenuItem className={css.footerHelpRow}>
+      {/* Help Button */}
+      <SidebarMenu>
+        <SidebarMenuItem>
           <SidebarMenuButton
-            className={cn('h-9 min-w-0 flex-1 gap-3', css.sidebarInteractive, css.sidebarNavItem)}
+            size="lg"
+            className={cn(css.sidebarInteractive, css.footerHelp, css.sidebarNavItem)}
+            render={<a href={`${HELP_CENTER_URL}/en/`} target="_blank" rel="noopener noreferrer" />}
             data-testid="list-item-need-help"
-            onClick={handleHelpClick}
           >
             <Tooltip>
               <TooltipTrigger render={<div />} className="flex min-w-0 cursor-pointer items-center gap-3">
@@ -115,8 +92,6 @@ export const SidebarCommonFooter = ({ isSafeSidebar = false }: { isSafeSidebar?:
           </div>
         </SidebarMenuItem>
       </SidebarMenu>
-
-      <HelpMenu anchorEl={helpMenuAnchor} onClose={handleHelpMenuClose} />
     </SidebarFooter>
   )
 }
