+++ apps/web/src/features/spaces/components/Sidebar/SidebarSkeleton/SidebarSkeleton.tsx
@@ -2,7 +2,6 @@ import type { ReactElement } from 'react'
 import {
   Sidebar,
   SidebarContent,
-  SidebarFooter,
   SidebarGroup,
   SidebarGroupContent,
   SidebarHeader,
@@ -10,23 +9,18 @@ import {
   SidebarMenuItem,
 } from '@/components/ui/sidebar'
 import { SidebarTopBar } from '../SidebarTopBar'
-import css from '../styles.module.css'
 
 const SIDEBAR_CONTAINER_CLASSNAME = '!p-0 border-r-0 group-data-[side=left]:border-r-0'
 const SIDEBAR_INNER_CLASSNAME = 'rounded-none rounded-tr-[8px] rounded-br-[8px] shadow-none'
 
-const Pulse = ({ className }: { className: string }): ReactElement => (
-  <div className={`bg-sidebar-border animate-pulse ${className}`} />
-)
-
-const NavRow = (): ReactElement => (
+const SkeletonRow = ({ heightClass }: { heightClass?: string } = {}): ReactElement => (
   <SidebarMenuItem>
-    <div className="relative flex h-9 min-h-9 w-full items-center rounded-md p-2 group-data-[collapsible=icon]:justify-center group-data-[collapsible=icon]:p-2">
-      <div className="flex w-full items-center gap-3 group-data-[collapsible=icon]:hidden">
-        <Pulse className="size-4 shrink-0 rounded-md" />
-        <Pulse className="h-4 min-h-4 flex-1 rounded-md" />
-      </div>
-      <Pulse className="hidden size-8 shrink-0 rounded-md group-data-[collapsible=icon]:block" />
+    <div className="flex items-center gap-2 px-3 py-2 group-data-[collapsible=icon]:justify-center group-data-[collapsible=icon]:px-0">
+      <div
+        className={`rounded bg-sidebar-border animate-pulse w-48 group-data-[collapsible=icon]:w-8 group-data-[collapsible=icon]:h-8 ${
+          heightClass ?? 'h-12'
+        }`}
+      />
     </div>
   </SidebarMenuItem>
 )
@@ -43,53 +37,25 @@ export const SidebarSkeleton = ({ contained = false }: { contained?: boolean }):
     >
       <SidebarHeader>
         <SidebarTopBar />
+        <SkeletonRow heightClass="h-18" />
       </SidebarHeader>
 
       <SidebarContent>
-        <SidebarGroup className={`${css.sidebarGroup} mb-2`}>
+        <SidebarGroup>
           <SidebarGroupContent>
-            <Pulse className="h-9 w-full rounded-xs group-data-[collapsible=icon]:w-9" />
-          </SidebarGroupContent>
-        </SidebarGroup>
-
-        <SidebarGroup className={css.sidebarGroup}>
-          <SidebarGroupContent>
-            <SidebarMenu className="gap-0.5">
-              <NavRow />
-              <NavRow />
-              <NavRow />
-              <NavRow />
-              <NavRow />
-              <NavRow />
-            </SidebarMenu>
-          </SidebarGroupContent>
-        </SidebarGroup>
-
-        <SidebarGroup className={css.sidebarGroup}>
-          <div className="px-2 py-1 group-data-[collapsible=icon]:hidden">
-            <Pulse className="h-3 w-20 rounded-md" />
-          </div>
-          <SidebarGroupContent>
-            <SidebarMenu className="gap-0">
-              <NavRow />
-              <NavRow />
-              <NavRow />
-              <NavRow />
+            <SidebarMenu>
+              <SkeletonRow />
+              <SkeletonRow />
+              <SkeletonRow />
+              <SkeletonRow />
+              <hr className="my-2 w-48 animate-pulse mx-auto" />
+              <SkeletonRow />
+              <SkeletonRow />
+              <SkeletonRow />
             </SidebarMenu>
           </SidebarGroupContent>
         </SidebarGroup>
       </SidebarContent>
-
-      <SidebarFooter>
-        <SidebarMenu className="gap-0.5">
-          <NavRow />
-          <NavRow />
-        </SidebarMenu>
-        <div className="flex items-center gap-3 px-3 py-2">
-          <Pulse className="size-5 rounded-full" />
-          <Pulse className="h-4 w-24 rounded-md group-data-[collapsible=icon]:hidden" />
-        </div>
-      </SidebarFooter>
     </Sidebar>
   )
 }
