+++ apps/web/src/features/spaces/components/Sidebar/SidebarSkeleton/index.ts
@@ -1,0 +1,1 @@
+export { SidebarSkeleton } from './SidebarSkeleton'
